+++ examples/data-sources/meraki_appliance_port/data-source.tf
@@ -1,0 +1,4 @@
+data "meraki_appliance_port" "example" {
+  network_id = "L_123456"
+  port_id    = "12"
+}