+++ examples/data-sources/meraki_appliance_vlans_settings/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_vlans_settings" "example" {
+  network_id = "L_123456"
+}