+++ examples/resources/meraki_appliance_port/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_port.example "<network_id>,<port_id>"