+++ examples/resources/meraki_appliance_vlans_settings/resource.tf
@@ -1,0 +1,4 @@
+resource "meraki_appliance_vlans_settings" "example" {
+  network_id    = "L_123456"
+  vlans_enabled = true
+}