+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Add `meraki_appliance_one_to_one_nat_rules` resource and data source
 - Add `meraki_appliance_port_forwarding_rules` resource and data source
 - Add `meraki_appliance_firewall_settings` resource and data source
+- Add `meraki_appliance_port` resource and data source
+- Add `meraki_appliance_vlans_settings` resource and data source
 
 ## 0.1.1